+++ services/api/src/trpc/app-router.generated.ts
@@ -5,10 +5,12 @@ import { router } from "@monark/common/trpc"
 
 import { authRouter } from "@monark/auth/server"
 import { featureFlagsRouter } from "@monark/feature-flags/server"
+import { usersRouter } from "@monark/users/server"
 
 export const appRouter = router({
   auth        : authRouter,
   featureFlags: featureFlagsRouter,
+  users       : usersRouter,
 })
 
 export type AppRouter = typeof appRouter